+++ modules/core/src/main/java/net/ashald/envfile/providers/dotenv/DotEnvFileParserFactory.java
@@ -8,8 +8,8 @@ public class DotEnvFileParserFactory implements EnvVarsProviderFactory {
 
     @Override
     public @NotNull
-    EnvVarsProvider createProvider() {
-        return new DotEnvFileParser();
+    EnvVarsProvider createProvider(boolean shouldSubstituteEnvVar) {
+        return new DotEnvFileParser(shouldSubstituteEnvVar);
     }
 
     @Override