+++ modules/core/src/main/java/net/ashald/envfile/providers/runconfig/RunConfigEnvVarsProvider.java
@@ -1,25 +1,26 @@
 package net.ashald.envfile.providers.runconfig;
 
-import net.ashald.envfile.EnvVarsProvider;
+import net.ashald.envfile.AbstractEnvVarsProvider;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class RunConfigEnvVarsProvider implements EnvVarsProvider {
+public class RunConfigEnvVarsProvider extends AbstractEnvVarsProvider {
+
+    public RunConfigEnvVarsProvider(boolean shouldSubstituteEnvVar) {
+        super(shouldSubstituteEnvVar);
+    }
+
+    @NotNull
+    @Override
+    protected Map<String, String> getEnvVars(@NotNull Map<String, String> runConfigEnv, String path) {
+        return new HashMap<>(runConfigEnv);
+    }
 
     @Override
     public boolean isEditable() {
         return false;
     }
 
-    @NotNull
-    @Override
-    public Map<String, String> process(@NotNull Map<String, String> runConfigEnv, String path, @NotNull Map<String, String> aggregatedEnv) {
-        Map<String, String> result = new HashMap<>(aggregatedEnv);
-
-        result.putAll(runConfigEnv);
-
-        return result;
-    }
 }