+++ modules/core/src/main/java/net/ashald/envfile/providers/yaml/YamlEnvFileParserFactory.java
@@ -8,8 +8,8 @@ public class YamlEnvFileParserFactory implements EnvVarsProviderFactory {
 
     @Override
     public @NotNull
-    EnvVarsProvider createProvider() {
-        return new YamlEnvFileParser();
+    EnvVarsProvider createProvider(boolean shouldSubstituteEnvVar) {
+        return new YamlEnvFileParser(shouldSubstituteEnvVar);
     }
 
     @NotNull