+++ modules/platform/src/main/java/net/ashald/envfile/platform/EnvFileSettings.java
@@ -6,15 +6,21 @@ import java.util.List;
 public class EnvFileSettings {
 
     private final boolean pluginEnabled;
+    private final boolean envVarsSubstitutionEnabled;
     private final List<EnvFileEntry> entries;
 
-    public EnvFileSettings(boolean isEnabled, List<EnvFileEntry> envFileEntries) {
+    public EnvFileSettings(boolean isEnabled, boolean substituteVars, List<EnvFileEntry> envFileEntries) {
         pluginEnabled = isEnabled;
+        envVarsSubstitutionEnabled = substituteVars;
         entries = envFileEntries;
     }
 
     public boolean isEnabled() {
         return pluginEnabled;
+    }
+
+    public boolean isSubstituteEnvVarsEnabled() {
+        return envVarsSubstitutionEnabled;
     }
 
     public List<EnvFileEntry> getEntries() {